update for training

+++ tools/demo.py
@@ -20,8 +20,10 @@ torch.set_num_threads(1)
 parser = argparse.ArgumentParser(description='tracking demo')
 parser.add_argument('--config', type=str, help='config file')
 parser.add_argument('--snapshot', type=str, help='model name')
-parser.add_argument('--video_name', default='', type=str, help='videos or image files')
+parser.add_argument('--video_name', default='', type=str,
+                    help='videos or image files')
 args = parser.parse_args()
+
 
 def get_frames(video_name):
     if not video_name:
@@ -46,7 +48,8 @@ def get_frames(video_name):
                 break
     else:
         images = glob(os.path.join(video_name, '*.jp*'))
-        images = sorted(images, key=lambda x: int(x.split('/')[-1].split('.')[0]))
+        images = sorted(images,
+                        key=lambda x: int(x.split('/')[-1].split('.')[0]))
         for img in images:
             frame = cv2.imread(img)
             yield frame
@@ -87,16 +90,20 @@ def main():
             outputs = tracker.track(frame)
             if 'polygon' in outputs:
                 polygon = np.array(outputs['polygon']).astype(np.int32)
-                cv2.polylines(frame, [polygon.reshape((-1,1,2))], True, (0,255,0),3)
-                mask = ((outputs['mask'] > cfg.TRACK.MASK_THERSHOLD) * 255).astype(np.uint8)
+                cv2.polylines(frame, [polygon.reshape((-1, 1, 2))],
+                              True, (0, 255, 0), 3)
+                mask = ((outputs['mask'] > cfg.TRACK.MASK_THERSHOLD) * 255)
+                mask = mask.astype(np.uint8)
                 mask = np.stack([mask, mask*255, mask]).transpose(1, 2, 0)
                 frame = cv2.addWeighted(frame, 0.77, mask, 0.23, -1)
             else:
                 bbox = list(map(int, outputs['bbox']))
-                cv2.rectangle(frame, (bbox[0],bbox[1]),
-                        (bbox[0]+bbox[2], bbox[1]+bbox[3]), (0,255,0),3)
+                cv2.rectangle(frame, (bbox[0], bbox[1]),
+                              (bbox[0]+bbox[2], bbox[1]+bbox[3]),
+                              (0, 255, 0), 3)
             cv2.imshow(video_name, frame)
             cv2.waitKey(40)
+
 
 if __name__ == '__main__':
     main()
